api_gateway/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
httpx  
cryptography
python-jose
pytest



